Add support for the sq_inplace_concat and sq_inplace_repeat call
git-svn-id: http://svn.code.sf.net/p/cxx/code@355 cb2f5747-f50d-0410-b093-b27f9c363ba5

+++ trunk/CXX/Demo/Python3/range.hxx
@@ -52,7 +52,7 @@ public:
 
     Py_ssize_t length() const;
     int item( int i ) const;
-    range *slice( int i, int j ) const;
+    range *slice( Py_ssize_t i, Py_ssize_t j ) const;
     range *extend( int k ) const;
     std::string asString() const;
